Added references as a Handler condition
Worker distribution decision making is now based on reference count
first followed by the load of the Worker. This means that existence of a
Buffer/Object reference in a Worker is the highest priority. If a Buffer
doesn't exist in the final candidate the StorageObject argument will be
passed as undefined to the foreign Task.